+++ R/req-body.R
@@ -10,6 +10,9 @@
 #' Adding a body to a request will automatically switch the method to POST.
 #'
 #' @inheritParams req_perform
+#' @param type Content type. For `req_body_file()`, the default
+#'  will will attempt to guess from the extension of `path`. For
+#'  `req_body_json()` the default is `application/json`.
 #' @returns A modified HTTP [request].
 #' @examples
 #' req <- request(example_url()) %>%
@@ -62,8 +65,6 @@ req_body_raw <- function(req, body, type = NULL) {
 #' @export
 #' @rdname req_body
 #' @param path Path to file to upload.
-#' @param type Content type. For `req_body_file()`, the default
-#'  will will attempt to guess from the extension of `path`.
 req_body_file <- function(req, path, type = NULL) {
   check_request(req)
   if (!file.exists(path)) {
@@ -86,9 +87,12 @@ req_body_json <- function(req, data,
                           auto_unbox = TRUE,
                           digits = 22,
                           null = "null",
+                          type = "application/json",
                           ...) {
   check_request(req)
   check_installed("jsonlite")
+  check_string(type)
+  check_content_type(type, "application/json", "json")
 
   params <- list2(
     auto_unbox = auto_unbox,
@@ -96,7 +100,7 @@ req_body_json <- function(req, data,
     null = null,
     ...
   )
-  req_body(req, data = data, type = "json", params = params)
+  req_body(req, data = data, type = "json", content_type = type, params = params)
 }
 
 #' @export
@@ -113,7 +117,12 @@ req_body_form <- function(.req, ...) {
   check_request(.req)
 
   data <- modify_body_data(.req$body$data, ...)
-  req_body(.req, data = data, type = "form")
+  req_body(
+    .req,
+    data = data,
+    type = "form",
+    content_type = "application/x-www-form-urlencoded"
+  )
 }
 
 #' @export
@@ -123,7 +132,12 @@ req_body_multipart <- function(.req, ...) {
 
   data <- modify_body_data(.req$body$data, ...)
   # data must be character, raw, curl::form_file, or curl::form_data
-  req_body(.req, data = data, type = "multipart")
+  req_body(
+    .req,
+    data = data,
+    type = "multipart",
+    content_type = NULL
+  )
 }
 
 modify_body_data <- function(data, ..., error_call = caller_env()) {
@@ -138,7 +152,7 @@ modify_body_data <- function(data, ..., error_call = caller_env()) {
 
 # General structure -------------------------------------------------------
 
-req_body <- function(req, data, type = NULL, params = list(), content_type = NULL) {
+req_body <- function(req, data, type, content_type, params = list()) {
   req$body <- list(
     data = data,
     type = type,
@@ -175,15 +189,6 @@ req_body_apply <- function(req) {
   data <- req$body$data
   type <- req$body$type
 
-  # Respect existing Content-Type if set
-  type_idx <- match("content-type", tolower(names(req$headers)))
-  if (!is.na(type_idx)) {
-    content_type <- req$headers[[type_idx]]
-    req$headers <- req$headers[-type_idx]
-  } else {
-    content_type <- req$body$content_type
-  }
-
   if (type == "raw-file") {
     size <- file.info(data)$size
     con <- file(data, "rb")
@@ -216,42 +221,29 @@ req_body_apply <- function(req) {
   } else if (type == "raw") {
     req <- req_body_apply_raw(req, data)
   } else if (type == "json") {
-    content_type <- check_req_content_type(content_type, "application/json", "json")
     json <- exec(jsonlite::toJSON, data, !!!req$body$params)
     req <- req_body_apply_raw(req, json)
   } else if (type == "multipart") {
     data <- unobfuscate(data)
-    content_type <- NULL
     req$fields <- data
   } else if (type == "form") {
     data <- unobfuscate(data)
-    content_type <- "application/x-www-form-urlencoded"
     req <- req_body_apply_raw(req, query_build(data))
   } else {
     abort("Unsupported request body `type`", .internal = TRUE)
   }
 
-  # Must set header afterwards
-  req <- req_headers(req, `Content-Type` = content_type)
-  req
-}
-
-check_req_content_type <- function(content_type,
-                                   valid_types,
-                                   valid_suffix = NULL,
-                                   call = caller_env()) {
-  if (is.null(content_type)) {
-    return(valid_types[[1]])
+  # Respect existing Content-Type if set
+  type_idx <- match("content-type", tolower(names(req$headers)))
+  if (!is.na(type_idx)) {
+    content_type <- req$headers[[type_idx]]
+    req$headers <- req$headers[-type_idx]
+  } else {
+    content_type <- req$body$content_type
   }
+  req <- req_headers(req, `Content-Type` = content_type)
 
-  check_content_type(
-    content_type,
-    valid_types = valid_types,
-    valid_suffix = valid_suffix,
-    inform_check_type = FALSE,
-    call = call
-  )
-  content_type
+  req
 }
 
 req_body_apply_raw <- function(req, body) {