+++ tests/testthat/test-req-body.R
@@ -58,8 +58,7 @@ test_that("can send any type of object as json", {
 
 test_that("can use custom json type", {
   resp <- request_test("/post") %>%
-    req_body_json(mtcars) %>%
-    req_headers(`Content-Type` = "application/ld+json") %>%
+    req_body_json(mtcars, type = "application/ld+json") %>%
     req_perform()
 
   expect_equal(
@@ -69,11 +68,10 @@ test_that("can use custom json type", {
 })
 
 test_that("non-json type errors", {
-  req <- request_test("/post") %>%
-    req_body_json(mtcars) %>%
-    req_headers(`Content-Type` = "application/xml")
-
-  expect_snapshot(req_perform(req), error = TRUE)
+  expect_snapshot(
+    req_body_json(request_test(), mtcars, type = "application/xml"),
+    error = TRUE
+  )
 })
 
 test_that("can send named elements as form/multipart", {